pyproject.toml
[project]
name = "pdfcards"
version = "0.1.0"
description = ""
authors = [
    {name = "Louis Wendler"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pypdf (>=6.0.0,<7.0.0)",
    "pdfplumber (>=0.11.7,<0.12.0)",
    "marker-pdf (>=1.8.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
pdfcards = "pdfcards.main:main"